package rename in the example app.
