+++ example/src/main/java/de/madcyph3r/example/tools/ActionBarButtons/FragmentActionBarButtons.java
@@ -1,14 +1,11 @@
-package de.madcyph3r.example.tools.oneHeadItemActionBarButtons;
+package de.madcyph3r.example.tools.ActionBarButtons;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBar;
 import android.view.Gravity;
 import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
